+++ django_perf_rec/api.py
@@ -10,7 +10,7 @@ from django.db import DEFAULT_DB_ALIAS
 
 from .cache import AllCacheRecorder
 from .db import AllDBRecorder
-from .utils import current_test
+from .utils import current_test, record_diff
 from .yaml import KVFile
 
 
@@ -112,7 +112,10 @@ class PerformanceRecorder(object):
         orig_record = self.records_file.get(self.record_name, None)
 
         if orig_record is not None:
-            assert self.record == orig_record, "Performance record did not match for {}".format(self.record_name)
+            assert self.record == orig_record, "Performance record did not match for {}\n{}".format(
+                self.record_name,
+                record_diff(orig_record, self.record)
+            )
 
         self.records_file.set_and_save(self.record_name, self.record)