+++ HISTORY.rst
@@ -8,6 +8,7 @@ Pending release
 
 * New release notes go here
 * Fix django session keys not being fingerprinted.
+* Show diff when records don't match.
 
 1.1.0 (2016-10-26)
 ------------------